backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
langchain==0.0.335
langchain-community==0.0.15
llama-cpp-python==0.2.20
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
neo4j==5.14.0
aiohttp==3.8.6
openai==1.2.4
python-multipart==0.0.6
scholarly==1.4.1
scikit-learn==1.3.2